Created new functions to move some code from main.cpp to other compilation units.
The compilation of the main.cpp file was taking to much memory. To solve this, we created
new compilation units to hold the code used in the parse and generation process.

Created the headers "parse_functions.hpp" and "generate_functions.hpp", also created
a ".cpp" file for each functions inside of those.

Created the headers "module_map.hpp", "parse_types.hpp" and "skipper.hpp" to share some common
types used in most of the parse/generation functions.

Added some include clauses to other headers that were not including all the necessary files.